Added file uploads to Assays to flex on the poor.

+++ INTEX/Models/Materials.cs
@@ -19,13 +19,13 @@ namespace INTEX.Models
 
         [Required(ErrorMessage = "Please the current Quantity Available")]
         [Display(Name = "Quanityt Available")]
-        public double QUANTITYAVAILABLE { get; set; }//set a reminder to replace this when below 100mg or something
+        public decimal QUANTITYAVAILABLE { get; set; }//set a reminder to replace this when below 100mg or something
 
         [Required(ErrorMessage = "Please select whether the material is reusable (Microscope, Petri Dishes, etc)")]
         [Display(Name = "Reusable")]
         public int REUSABLE { get; set; }//like do we need to replace the substance?  
 
         [Display(Name = "Cost Per Miligram (if not reusable)")]
-        public double COST { get; set; }//cost per mg
+        public decimal COST { get; set; }//cost per mg
     }
 }
